move topk heap up to TimeSeriesShard
